Rename clusterstatus to clusterstate
